move lib/auth/predicate to lib/predicate for dependency referencing reasons
