+++ lib/predicate/builtins.go
