+++ lib/predicate/partial/builtins.go
