+++ lib/predicate/partial/lower.go
