+++ lib/predicate/partial/solver.go
