+++ lib/predicate/partial/solver_test.go
