+++ lib/predicate/predicate.go
